// src/types/shared.ts
export type FetchParams = {
  page: number;
  limit: number;
  sortKey: string | null;
  sortOrder: "asc" | "desc" | "" | null;
  searchQuery: null | {
    value: string | null;
    fields: string[];
  };
  query: null | Record<string, MappedFilters[]>;
  fields?: DynamicFilter[];
};

export type DataSource<
  T extends GenericObject = GenericObject,
  Remote extends boolean = boolean
> = Remote extends true
  ? (params: FetchParams) => MaybePromise<{
      docs: T[];
      totalDocs: number;
      totalPages: number;
    }>
  : () => MaybePromise<T[]>;

export type ActionParams<
  T extends GenericObject = GenericObject,
  KeyPath = any
> = {
  nbSelected: number;
  selectAll: boolean;
  selected: T[];
  fetchParams: FetchParams;
  tableApi: DataApi<T, KeyPath>;
};

export type Action<
  T extends GenericObject = GenericObject,
  KeyPath = any,
  TParams extends ActionParams<T, KeyPath> = ActionParams<T, KeyPath>
> = {
  label: string;
  icon: string;
  action?: (actionParams: TParams) => void;
  link?: string | AppTypes["routeLocation"];
  permissions?: (AppTypes["permissionKey"] | AppTypes["permissionKey"][])[];
  condition?: (data: T[], params: TParams) => boolean;
};

export type DataApi<T = GenericObject, KeyPath = any> = {
  refreshData: () => void;
  setSearchQuery: (value: string) => void;
  resetFilters: () => void;
  setSort: (key: KeyPath, dir?: "asc" | "desc" | null) => void;
  setPage: (page: number) => void;
  setPageSize: (size: number) => void;
  updateRow: (selector: (row: T) => boolean, updater: (row: T) => T) => boolean;
  updateRows: (
    selector: (row: T) => boolean,
    updater: (row: T) => T
  ) => boolean;
};
// ...
export type FullQueryState = ReturnType<typeof useQueryState>;